requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
discord.py==2.2.2
frozenlist==1.3.3
multidict==6.0.4
openai==0.27.4
python-dotenv==1.0.0
tqdm==4.65.0
yarl==1.9.1
